allow a `format` and explicit Buffer to be used
Also adds new `flush()` function, and `format` getter.

This allows a specific Buffer instance to be the "backing store"
for a Canvas instance. This is especially useful when interfacing
directly with a Linux framebuffer, as it allows for faster
rendering and the `format` can be specified to match what the
framebuffer is expecting.

For example, the PiTFT screen expects 16-bits per pixel to
be written to `/dev/fb1`, so we can specify that the
CAIRO_FORMAT_RGB16_565 format be used for the Canvas instance.

Overall this API could be bikeshedded a bit perhaps, but
it's functional at least.

Here's a couple videos from my Raspberry Pi:

 - https://cloudup.com/iz--PO83YR2
 - https://cloudup.com/iJmL2M-mdw3