Render SVG img elements when librsvg is available
In the browser, <canvas> can render SVG <img> elements.

This commit enables node-canvas to do the same, if librsvg is installed.
Note that librsvg remains an optional dependency.

Use cases for embedded SVG images include the rendering of d3 charts,
which often blend SVG and canvas content.

Big thanks to petarov and scott113341 for much of the original code.