Backends support
This is a stripped-down version of pull-request #571 focused only on the
backends support, but adding also support for PDF and SVG backends and
passing the tests. it also retain backwards compatibility with the old API,
so this would be easy to merge since it's mostly to move around the PDF and
SVG code and checks to independent classes. There's some code for
`pdfStream` and similar methods that I think they should be moved too since
they are mostly backend specific, and make more sense there. I didn't do
that since it would break the API, but maybe it would be added some
functions on the Canvas object that proxy the petitions to the backend to
retain compatibility.

I know it's a big pull-request, so I'm open for comments over it.