Use std::vector instead of toBuffer.cc; use our own jpeg mem dest
free(data) was omitted from the code because it caused a crash with JPEG
formats. That turns out to be because at least on Windows using prebuilt
libjpeg-turbo, the malloc() that jpeg_mem_dest uses doesn't work with
the free() that MSVC provides (different heap). Since libjpeg doesn't
provide a way to call its particular free() function, we need to use our
own version of jpeg_mem_dest.

That omission of free() also caused a leak for toBuffer("image/png").

Replaced the toBuffer code with std::vector.