remove backends
This is too much indirection. Backends have led to a few bugs by
being confusing, and there is zero benefit to such an abstraction.
It's only added LOC and more jumping around files to figure out
what's happening.

IIRC backends were added with the intent to have framebuffer
backends but that never got merged because it's way out of scope.

I broke the JS ctors in ce29f697ced288b8d948e92b93b91d32ca3353d5
so we got lucky and don't have to call this a breaking change.